GUI Issue #75 - Moving config parsing and client creation from broker to server; bug fixes
