+++ ait/server/handlers/ait_packet_handler.py
@@ -1,20 +1,20 @@
 from ait.core import log, tlm
+from handler import Handler
 
 
-class AITPacketHandler(object):
+class AitPacketHandler(Handler):
+
     def __init__(self, input_type=None, output_type=None, **kwargs):
-        super(AITPacketHandler, self).__init__(input_type, output_type)
+        super(AitPacketHandler, self).__init__(input_type, output_type)
         self.packet = kwargs.get('packet', None)
 
         if not self.packet:
-            msg = 'AITPacketHandler: No packet name provided in handler config as key "name"'
-            log.error(msg)
+            msg = 'AITPacketHandler: No packet name provided in handler config as key "packet"'
             raise ValueError(msg)
 
         tlm_dict = tlm.getDefaultDict()
         if self.packet not in tlm_dict:
             msg = 'AITPacketHandler: Packet name {} not present in telemetry dictionary'.format(self.packet)
-            log.error(msg)
             raise ValueError(msg)
 
         self._pkt_defn = tlm_dict[self.packet]
